/editmaps changes
lots of changes to finalise it

- compress image after saving image
- make image_util for compress function
- move the command into admin commands
